refactor: correct use of typescript
